requirements.txt
streamlit==1.34.0
streamlit_extras
groq
python-docx
docx2txt
pypdf
